Update Facebook webchat analyzer for Bro 2.x.
This revamped version fo facebook.bro uses the new bodies.bro script to keep
track of HTTP bodies. The previous version had a tightly couples implementation
of the two rather disjoint concepts. Now, facebook.bro is much cleaner and
focuses only on the tasks it needs to get done, which is parsing HTTP bodies.

With the new Bro 2.x logging framework, the output of this script is now a
separate stream that logs into facebook.log. Here is some example output:

    #separator \x09
    #path   facebook
    #fields timestamp       chat_from       chat_to chat_msg
    #types  string  string  string  string
    1303218454567   Mondo Cheeze    Udder Kaos      So I need the URL, dude.  What is it?
    1303218465938   Udder Kaos      Mondo Cheeze    the URL?
    1303218474259   Mondo Cheeze    Udder Kaos      Yeah for the secret image
    1303218481721   Udder Kaos      Mondo Cheeze    ok lemme see
    ...