Add a rough Duqu detector.
This script is based on Symantec's description of Duqu's network behavior, as
described in http://bit.ly/duqu-analysis. It implements a small state machine
that tracks Duqu's HTTP request/response behavior. Currently, it does not track
the SMB C&C behavior. If you have a sample trace of Duqu, please contact me so
that I can improve it.