Improve the Duqu detector.
The previous version of the detector incorrectly handled the http_request event
and tried to use yet unpopulated fields of c$http in there, such as
request_body_len or the cookie value. Now we process the http message in the
event http_message_done to make sure that the message contains all the
necessary fields.

Based on the log files from CrySys Lab, it seems that the POST occurs before
the initial JPEG GET, as reported in the Symantec report. However, this could
just be a capturing artefact, which is why the detector's state machine still
uses the JPEG GET request as prerequisite event before the POST. If other
traces confirm that the GET and POST are decoupled, it makes sense to decouple
the state machine.